implemented framing on client and server
messages without framing of length, would constantly get mixed up, sent incomplete, etc, so an 8 byte prefix to the buffer sent was introduced to indicate how long both clients and servers expect messages to be when declaring it a full "message" and not just whatever they get

could maybe be optimized? but i think this is perfectly acceptable and does it's job well. tested with various frame rates, all works wonderfully.

I think this project is MOSTLY complete, could look into handling disconnections, so that might be the last commit